pyproject.toml
[tool.poetry]
name = "workflow"
version = "0.1.0"
description = ""
authors = ["skyline93 <greene9832@163.com>"]
readme = "README.md"
packages = [{include = "workflow"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
psycopg2-binary = "^2.9.5"
toml = "^0.10.2"
pydantic = "^1.10.4"


[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
secondary = false

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"
pytest = "^7.2.0"
ipython = "^8.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
